requirements.txt
aspose-words==23.12.0
openai
python-dotenv
streamlit
